+++ common/src/main/java/io/github/how_bout_no/outvoted/Outvoted.java
@@ -46,10 +46,11 @@ public class Outvoted {
         GeckoLib.initialize();
         GeckoLibMod.DISABLE_IN_DEV = true;
 
-        ModEntityTypes.ENTITY_TYPES.register();
-        ModItems.ITEMS.register();
         ModBlocks.BLOCKS.register();
         ModBlocks.BLOCK_ITEMS.register();
+        ModEntityTypes.ENTITY_TYPES.register();
+        ModItems.ITEMS.register();
+        ModLootConditionTypes.CONDITIONS.register();
         ModRecipes.RECIPES.register();
         ModSounds.SOUNDS.register();
         ModTags.init();